+++ tests/apply.test.js
@@ -1691,310 +1691,3 @@ it('should not replace multiple instances of the same class in a single selector
     }
   `)
 })
-
-it('does not warn for simple utilities', async () => {
-  let config = {
-    content: [{ raw: html`<div class="bar"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    @layer utilities {
-      .foo {
-        color: blue;
-      }
-    }
-    .bar {
-      @apply foo;
-    }
-  `
-
-  await doesNotShowWarnings(run(input, config))
-})
-
-it('does not warn for simple utilities with simple variants', async () => {
-  let config = {
-    content: [{ raw: html`<div class="bar"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    @layer utilities {
-      .foo {
-        color: blue;
-      }
-    }
-    .bar {
-      @apply hover:foo;
-    }
-  `
-
-  await doesNotShowWarnings(run(input, config))
-})
-
-it('does not warn for simple utilities with complex variants', async () => {
-  let config = {
-    content: [{ raw: html`<div class="bar"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    @layer utilities {
-      .foo {
-        color: blue;
-      }
-    }
-    .bar {
-      @apply group-hover:foo;
-    }
-  `
-
-  await doesNotShowWarnings(run(input, config))
-})
-
-it('does not warn when matching rules with multiple simple selectors without multiple classes', async () => {
-  let config = {
-    content: [{ raw: html`<div class="bar"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    @layer utilities {
-      header .foo {
-        color: blue;
-      }
-    }
-    .bar {
-      @apply foo;
-    }
-  `
-
-  await doesNotShowWarnings(run(input, config))
-})
-
-it('does not warn when using pseudo elements in a utility', async () => {
-  let config = {
-    content: [{ raw: html`<div class="bar"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    .bar {
-      @apply placeholder-blue-600;
-    }
-  `
-
-  await doesNotShowWarnings(run(input, config))
-})
-
-it('does not warn when using pseudo elements from variants', async () => {
-  let config = {
-    content: [{ raw: html`<div class="bar"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    .bar {
-      @apply placeholder:text-blue-600;
-    }
-  `
-
-  await doesNotShowWarnings(run(input, config))
-})
-
-it('does not warn when using built in utilities that have multiple selector components', async () => {
-  let config = {
-    content: [{ raw: html`<div class="bar baz"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    .bar {
-      @apply divide-x-4;
-    }
-    .baz {
-      @apply space-x-4;
-    }
-  `
-
-  await doesNotShowWarnings(run(input, config))
-})
-
-it('warns when matching multiple rules', async () => {
-  let config = {
-    content: [{ raw: html`<div class="bar"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    @layer utilities {
-      .foo {
-        color: blue;
-      }
-      .foo {
-        background: green;
-      }
-    }
-    .bar {
-      @apply foo;
-    }
-  `
-
-  await showsWarnings(run(input, config), [/matching multiple rules/])
-})
-
-it('warns when matching rules that extend built in utilities', async () => {
-  let config = {
-    content: [{ raw: html`<div class="bar"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    @layer utilities {
-      header .bg-red-500 {
-        text-decoration: underline;
-      }
-    }
-    .bar {
-      @apply bg-red-500;
-    }
-  `
-
-  await showsWarnings(run(input, config), [/matching multiple rules/])
-})
-
-it('warns when matching rules with multiple of the same class', async () => {
-  let config = {
-    content: [{ raw: html`<div class="bar"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    @layer utilities {
-      .foo .foo {
-        color: blue;
-      }
-    }
-    .bar {
-      @apply foo;
-    }
-  `
-
-  await showsWarnings(run(input, config), [/with multiple classes/])
-})
-
-it('warns when matching rules with multiple different classes', async () => {
-  let config = {
-    content: [{ raw: html`<div class="bar"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    @layer utilities {
-      .foo .baz {
-        color: blue;
-      }
-    }
-    .bar {
-      @apply foo;
-    }
-  `
-
-  await showsWarnings(run(input, config), [/with multiple classes/])
-})
-
-it('warns when matching rules with multiple selectors', async () => {
-  let config = {
-    content: [{ raw: html`<div class="bar"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    @layer utilities {
-      .foo,
-      .baz {
-        color: blue;
-      }
-    }
-    .bar {
-      @apply foo;
-    }
-  `
-
-  await showsWarnings(run(input, config), [/with multiple selectors/])
-})
-
-it('warns when using multiple classes when one is inside :has()', async () => {
-  let config = {
-    content: [{ raw: html`<div class="bar"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    @layer utilities {
-      .foo:has(.baz) {
-        color: blue;
-      }
-    }
-    .bar {
-      @apply foo;
-    }
-  `
-
-  await showsWarnings(run(input, config), [/with multiple classes/])
-})
-
-it('warns when nested applies are encountered', async () => {
-  let config = {
-    content: [{ raw: html`<div class="qux"></div>` }],
-    plugins: [],
-  }
-
-  let input = css`
-    @tailwind utilities;
-    @layer utilities {
-      .foo {
-        color: red;
-      }
-      .bar {
-        @apply foo;
-      }
-      .baz {
-        @apply bar;
-      }
-    }
-    .qux {
-      @apply baz;
-    }
-  `
-
-  await showsWarnings(run(input, config), [/nested expansion/, /nested expansion/])
-})
-
-async function showsWarnings(promise, expectedWarnings) {
-  let result = await promise
-  let warnings = result.messages.filter((msg) => msg.type === 'warning').map((msg) => msg.text)
-
-  expect(warnings).toHaveLength(expectedWarnings.length)
-
-  for (let [index, actual] of warnings.entries()) {
-    expect(actual).toMatch(expectedWarnings[index])
-  }
-}
-
-async function doesNotShowWarnings(promise) {
-  await showsWarnings(promise, [])
-}
